modifica salvataggio dati libri e autori nel database

+++ csharp-biblioteca-db/Biblioteca.cs
@@ -43,7 +43,7 @@ namespace csharp_biblioteca_db
                 db.scaffaleAdd(nuovo.Numero);
         }
 
-        public void AggiungiLibro(int codice , string sTitolo , int iAnno , string sSettore , int iNumeroPagine ,string sScaffale,List<Autore>listaAutori)
+        public void AggiungiLibro(long codice , string sTitolo , int iAnno , string sSettore , int iNumeroPagine ,string sScaffale,List<Autore>listaAutori)
         {
             Libro nuovoLibro = new Libro(codice ,sTitolo ,iAnno ,sSettore , iNumeroPagine ,sScaffale);
             nuovoLibro.Stato = Stato.Disponibile;
